+++ src/__tests__/__mocks__/responses/programmingLanguages.js
@@ -1,0 +1,41 @@
+/**
+ * The answer to the /programming-languages call
+ */
+export const items = [
+  {
+    id: 1,
+    name: "PHP",
+    view_order: 1,
+    knowledge_percentage: 95,
+    created_at: "2023-02-02T20:06:03.000000Z",
+    updated_at: "2023-02-02T20:06:03.000000Z",
+    deleted_at: null
+  },
+  {
+    id: 2,
+    name: "Java",
+    view_order: 2,
+    knowledge_percentage: 75,
+    created_at: "2023-02-02T20:06:14.000000Z",
+    updated_at: "2023-02-02T20:06:14.000000Z",
+    deleted_at: null
+  },
+  {
+    id: 3,
+    name: "Javascript",
+    view_order: 3,
+    knowledge_percentage: 85,
+    created_at: "2023-02-02T20:06:26.000000Z",
+    updated_at: "2023-02-02T20:06:26.000000Z",
+    deleted_at: null
+  },
+  {
+    id: 4,
+    name: "Python",
+    view_order: 4,
+    knowledge_percentage: 60,
+    created_at: "2023-02-02T20:06:38.000000Z",
+    updated_at: "2023-02-02T20:06:38.000000Z",
+    deleted_at: null
+  },
+];